Potential Fix for Autonomous skipping steps
We have been chasing a bug that has the autonomous opmodes skipping steps.  We went back and compared the last time any autonomous opmode worked (this was meet 2).  There were two differences: 1) we turned the drive.java debug off. 2) the chosing of a path/array was changed to 0,1,4.
After some thinking and a quick test on a mule bot, it appears that turning thedrive.java debug back on or reverting back to 4,1,0 make automodes more reliable. We still need to validate this in the shop on the Onion Ring Factory.